Rails with js

+++ app/controllers/items_controller.rb
@@ -5,8 +5,10 @@ class ItemsController < ApplicationController
     if params[:restaurant_id]
       @items = Restaurant.find(params[:restaurant_id]).items
       @restaurant = Restaurant.find(params[:restaurant_id])
-    else
-      @items = Item.all
+      respond_to do |format|
+        format.html {render :index}
+        format.json {render json: @items}
+      end
     end
   end
 
@@ -24,7 +26,7 @@ class ItemsController < ApplicationController
   def create
     @item = Item.new(item_params)
     if @item.save
-      redirect_to restaurant_item_path(@item.restaurant_id, @item)
+      render json: @item, status: 201
     else
       @restaurant = Restaurant.find(params[:item][:restaurant_id])
       render :new
@@ -35,9 +37,10 @@ class ItemsController < ApplicationController
     if params[:restaurant_id]
       @item = Restaurant.find(params[:restaurant_id]).items.find(params[:id])
       @restaurant = Restaurant.find(params[:restaurant_id])
-    else
-      @item = Item.find(params[:id])
-      @restaurant = Restaurant.find(params[:restaurant_id])
+      respond_to do |format|
+        format.html {render :show}
+        format.json {render json: @item}
+      end
     end
   end
 
@@ -51,22 +54,27 @@ class ItemsController < ApplicationController
       redirect_to restaurant_items_path(@restaurant)
     end
   end
+
   def update
     @item = Item.find(params[:id])
 
     if @item.update(item_params)
-      redirect_to restaurant_item_path(@item.restaurant_id, @item)
+        redirect_to restaurant_path(@item.restaurant_id)
     else
-      @restaurant = Restaurant.find(params[:item][:restaurant_id])
+      # @restaurant = Restaurant.find(params[:item][:restaurant_id])
       render :edit
     end
   end
 
   def destroy
-    item = Item.find(params[:id])
-    restaurant = Restaurant.find_by(id: item.restaurant_id)
-    item.destroy
-    redirect_to restaurant_items_path(restaurant)
+    if current_user.restaurants.uniq.include?(Restaurant.find(params[:restaurant_id]))
+      item = Item.find(params[:id])
+      restaurant = Restaurant.find_by(id: item.restaurant_id)
+      item.destroy
+      redirect_to restaurant_path(restaurant)
+    else
+      redirect_to restaurants_path
+    end
   end
 
   private
